% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furly.R
\name{furly}
\alias{furly}
\title{Fast downloads and parsing of multiple JSON urls}
\usage{
furly(urls)
}
\arguments{
\item{urls}{a list of urls}
}
\value{
a list of parsed json objects
}
\description{
Fast downloads and parsing of multiple JSON urls
}
\examples{
urls <- paste0(
  "http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/rest/v4/hsapiens/feature/gene/ATM/snp?",
  "limit=", 500,
  "&skip=", c(-1, 500, 1000, 1500),
  "&skipCount=false&count=false&Output\%20format=json&merge=false"
)
res<-furly(urls)
}
