% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furly.R
\name{furly}
\alias{furly}
\title{Fast downloads and parsing of multiple JSON urls}
\usage{
furly(urls)
}
\arguments{
\item{urls}{a list of urls}
}
\value{
a list of parsed json objects
}
\description{
Fast downloads and parsing of multiple JSON urls
}
\examples{
url <- "https://bit.ly/cellbase1"
url2 <- "https://bit.ly/cellbase2"
url3 <- "https://bit.ly/cellbase3"
url4 <- "https://bit.ly/cellbase4"
urls <- list(url,url2,url3,url4)
res<-furly(urls)
}
